Added an offset table mode setting
Fragmented means that each data block pointed to in an offset table will be allocated in a fragmented fashion, so when looking at the resulting offset table, the offsets will not be in any meaningful order. In Contiguous mode, all data blocks are concatenated in one giant blob and allocated as a unit, so all offsets will end up being in order.

Fragmented mode has a higher chance of finding a successful allocation solution, while with Contiguous mode it's a bit easier to look at the resulting offset table in a hex editor, for example.