Horrible
Mother3Project.CreateNew is basically a helper that takes file paths and turns them into a new project and also decompiles everything. The FileSystemPath parameters aren't really useful from a user perspective, so it's been changed to use simple physical file paths. As a result, this method is now a garbage dump.